change all to double better precesion

+++ srcs/draw.c
@@ -49,18 +49,18 @@ void	setup(t_mlx *m)
 	m->ad = mlx_get_data_addr(m->img, &(m->b_p_p), &(m->s_l), &(m->end));
 }
 
-float	calc_cor(t_mlx *m, int cor, int md)
+double	calc_cor(t_mlx *m, int cor, int md)
 {
 	if (md == 0)
 		return (((m->xy_mnmx[0] + m->xy_mnmx[1]) / 2 +
-				(cor - (float)WIN_W / 2) * m->scale[0]));
+				(cor - (double)WIN_W / 2) * m->scale[0]));
 		return (((m->xy_mnmx[2] + m->xy_mnmx[3]) / 2 -
-				(cor - (float)WIN_H / 2) * m->scale[1]));
+				(cor - (double)WIN_H / 2) * m->scale[1]));
 }
 
-void	shift(int k, t_mlx *m, float scale)
+void	shift(int k, t_mlx *m, double scale)
 {
-	float	shift;
+	double	shift;
 
 	shift = WIN_W * scale * fabs(m->scale[0]);
 	if (k >= 125)
